+++ scripts/test/Selenium/Agent/AgentTicketActionCommonACL.t
@@ -1,0 +1,243 @@
+# --
+# Copyright (C) 2001-2016 OTRS AG, http://otrs.com/
+# --
+# This software comes with ABSOLUTELY NO WARRANTY. For details, see
+# the enclosed file COPYING for license information (AGPL). If you
+# did not receive this file, see http://www.gnu.org/licenses/agpl.txt.
+# --
+use strict;
+use warnings;
+use utf8;
+use vars (qw($Self));
+
+# note: this UT covers bug #11874 - Restrict service based on state when posting a note
+
+# get selenium object
+my $Selenium = $Kernel::OM->Get('Kernel::System::UnitTest::Selenium');
+
+$Selenium->RunTest(
+    sub {
+
+        $Kernel::OM->Get('Kernel::System::SysConfig')->ConfigItemUpdate(
+            Valid => 1,
+            Key   => 'Ticket::Service',
+            Value => 1,
+        );
+
+        # get helper object
+        $Kernel::OM->ObjectParamAdd(
+            'Kernel::System::UnitTest::Helper' => {
+                RestoreSystemConfiguration => 1,
+            },
+        );
+        my $Helper = $Kernel::OM->Get('Kernel::System::UnitTest::Helper');
+
+        # get ACL object
+        my $ACLObject = $Kernel::OM->Get('Kernel::System::ACL::DB::ACL');
+
+        # import test ACL
+        $ACLObject->ACLImport(
+            Content => <<"EOF",
+- ChangeBy: root\@localhost
+  ChangeTime: 2016-02-16 03:08:58
+  Comment: ''
+  ConfigChange:
+    Possible:
+      Ticket:
+        Service:
+        - UT Testservice 1
+  ConfigMatch:
+    Properties:
+      Ticket:
+        State:
+        - new
+  CreateBy: root\@localhost
+  CreateTime: 2016-02-16 02:55:35
+  Description: ''
+  ID: '1'
+  Name: ThisIsAUnitTestACL
+  StopAfterMatch: 0
+  ValidID: '1'
+EOF
+            OverwriteExistingEntities => 1,
+            UserID                    => 1,
+        );
+
+        # create test user and login
+        my $TestUserLogin = $Helper->TestUserCreate(
+            Groups => [ 'admin', 'users' ],
+        ) || die "Did not get test user";
+
+        $Selenium->Login(
+            Type     => 'Agent',
+            User     => $TestUserLogin,
+            Password => $TestUserLogin,
+        );
+
+        # get config object
+        my $ConfigObject = $Kernel::OM->Get('Kernel::Config');
+
+        # get script alias
+        my $ScriptAlias = $ConfigObject->Get('ScriptAlias');
+
+        # after login, we need to navigate to the ACL deployment to make the imported ACL work
+        $Selenium->VerifiedGet("${ScriptAlias}index.pl?Action=AdminACL;Subaction=ACLDeploy");
+        $Self->False(
+            index( $Selenium->get_page_source(),
+                'ACL information from database is not in sync with the system configuration, please deploy all ACLs.' )
+                > -1,
+            "ACL deployment successful."
+        );
+
+        # add a customer
+        my $CustomerUserLogin = $Kernel::OM->Get('Kernel::System::CustomerUser')->CustomerUserAdd(
+            UserFirstname  => 'Huber',
+            UserLastname   => 'Manfred',
+            UserCustomerID => 'A124',
+            UserLogin      => 'customeruser_' . $Helper->GetRandomID(),
+            UserPassword   => 'some-pass',
+            UserEmail      => $Helper->GetRandomID() . '@localhost.com',
+            ValidID        => 1,
+            UserID         => 1,
+        );
+
+        # get ticket object
+        my $TicketObject = $Kernel::OM->Get('Kernel::System::Ticket');
+
+        # create test ticket
+        my $TicketID = $TicketObject->TicketCreate(
+            Title        => 'Selenium Ticket',
+            Queue        => 'Raw',
+            Lock         => 'unlock',
+            Priority     => '3 normal',
+            State        => 'new',
+            CustomerID   => '123465',
+            CustomerUser => $CustomerUserLogin,
+            OwnerID      => 1,
+            UserID       => 1,
+        );
+        $Self->True(
+            $TicketID,
+            "TicketCreate - ID $TicketID",
+        );
+
+        # create some testservices
+        my $ServiceObject = $Kernel::OM->Get('Kernel::System::Service');
+
+        my $ServiceID;
+        my @Services;
+        for my $Count ( 1 .. 3 ) {
+            $ServiceID = $ServiceObject->ServiceAdd(
+                Name    => "UT Testservice $Count",
+                ValidID => 1,
+                UserID  => 1,
+            );
+            push @Services, $ServiceID;
+
+            $ServiceObject->CustomerUserServiceMemberAdd(
+                CustomerUserLogin => $CustomerUserLogin,
+                ServiceID         => $ServiceID,
+                Active            => 1,
+                UserID            => 1,
+            );
+
+            $Self->True(
+                $ServiceID,
+                "Test service $Count ($ServiceID) created and assigned to customer user",
+            );
+        }
+
+        # navigate to AgentTicketZoom screen of created test ticket
+        $Selenium->VerifiedGet("${ScriptAlias}index.pl?Action=AgentTicketZoom;TicketID=$TicketID");
+
+        # wait for displaying submenu items for 'People' ticket menu item
+        $Selenium->WaitFor(
+            JavaScript =>
+                'return typeof($) === "function" && $("#nav-Communication ul").css({ "height": "auto", "opacity": "100" });'
+        );
+
+        # click on 'Note' and switch window
+        $Selenium->find_element("//a[contains(\@href, \'Action=AgentTicketNote;TicketID=$TicketID' )]")->click();
+
+        $Selenium->WaitFor( WindowCount => 2 );
+        my $Handles = $Selenium->get_window_handles();
+        $Selenium->switch_to_window( $Handles->[1] );
+
+        # wait until page has loaded
+        $Selenium->WaitFor( JavaScript => 'return typeof($) === "function" && $("#DestQueueID").length' );
+
+        # check for entries in the service selection, there should be only one
+        $Self->True(
+            $Selenium->execute_script(
+                "return \$('#ServiceID option').length"
+            ),
+            "There is only one entry in the service selection",
+        );
+
+        # delete test ACL
+        my $ACLData = $ACLObject->ACLGet(
+            Name   => 'ThisIsAUnitTestACL',
+            UserID => 1,
+        );
+
+        my $Success = $ACLObject->ACLDelete(
+            ID     => $ACLData->{ID},
+            UserID => 1,
+        );
+        $Self->True(
+            $Success,
+            "ACL with ID $ACLData->{ID} is deleted"
+        );
+
+        # delete created test tickets
+        $Success = $TicketObject->TicketDelete(
+            TicketID => $TicketID,
+            UserID   => 1,
+        );
+        $Self->True(
+            $Success,
+            "Ticket with ticket ID $TicketID is deleted"
+        );
+
+        # deploy again after we deleted the test acl
+        $Selenium->VerifiedGet("${ScriptAlias}index.pl?Action=AdminACL;Subaction=ACLDeploy");
+        $Self->False(
+            index( $Selenium->get_page_source(),
+                'ACL information from database is not in sync with the system configuration, please deploy all ACLs.' )
+                > -1,
+            "ACL deployment successful."
+        );
+
+        # delete services and relations
+        my $DBObject = $Kernel::OM->Get('Kernel::System::DB');
+        $Success = $DBObject->Do(
+            SQL  => "DELETE FROM service_customer_user WHERE customer_user_login = ?",
+            Bind => [ \$CustomerUserLogin ],
+        );
+        $Self->True(
+            $Success,
+            "Deleted service relations for $CustomerUserLogin",
+        );
+        for my $ServiceID (@Services) {
+            $Success = $DBObject->Do(
+                SQL  => "DELETE FROM service WHERE ID = ?",
+                Bind => [ \$ServiceID ],
+            );
+            $Self->True(
+                $Success,
+                "Deleted service with ID $ServiceID",
+            );
+        }
+
+        $Success = $DBObject->Do(
+            SQL  => "DELETE FROM customer_user WHERE login = ?",
+            Bind => [ \$CustomerUserLogin ],
+        );
+        $Self->True(
+            $Success,
+            "Deleted Customer $CustomerUserLogin",
+        );
+    }
+);
+
+1;